Move development info to the wiki and update vagrant_setup to include selendium and chromium-browser installs.

+++ vagrant_setup.sh
@@ -1,21 +1,30 @@
 #!/bin/bash
 
-#apt-get update
+apt-get update
 
 function set_git_config_commit_vim {
   git config --global core.editor "vim"
 }
 
 function install {
+  echo "INSTALLING: $1"
   apt-get -y install $1 || { echo "ERROR: installing ${1}"; exit 1; }
 }
 
 function install_node_package_manager {
-  install npm
+  if [ ! -e /usr/bin/npm ]; then
+    install npm
+  fi
   #Switch to http version (as apposed to https) due to proxy auth issues.
   #https://github.com/npm/npm/issues/2119
   npm config set registry http://registry.npmjs.org/
   npm install npm -g.
+}
+
+#upgrade nodejs
+function install_nodejs {
+  curl -sL https://deb.nodesource.com/setup | sudo bash -
+  apt-get install -y nodejs
 }
 
 
@@ -35,5 +44,12 @@ install python-sqlalchemy
 install sqlite3
 install libsqlite3-dev
 
+
+install curl
+install_nodejs
+
 install_node_package_manager
 
+install xvfb
+pip install selenium
+install chromium-browser
